+++ app/admin/questions/[id]/page.tsx
@@ -1,0 +1,3 @@
+export function ManageUserPage() {
+  return null;
+}
